refactor(reborn): modularize conversation contracts
