+++ crates/ironclaw_conversations/src/types.rs
@@ -1,0 +1,125 @@
+use chrono::{DateTime, Utc};
+use ironclaw_host_api::{AgentId, ProjectId, TenantId, ThreadId, UserId};
+use ironclaw_turns::{
+    AcceptedMessageRef, ReplyTargetBindingRef, RunProfileRequest, SourceBindingRef,
+    SubmitTurnResponse, TurnActor, TurnScope,
+};
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    AdapterInstallationId, AdapterKind, ExternalActorRef, ExternalConversationRef, ExternalEventId,
+    InboundMessageContentRef,
+};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ResolveConversationRequest {
+    pub tenant_id: TenantId,
+    pub adapter_kind: AdapterKind,
+    pub adapter_installation_id: AdapterInstallationId,
+    pub external_actor_ref: ExternalActorRef,
+    pub external_conversation_ref: ExternalConversationRef,
+    pub external_event_id: ExternalEventId,
+    pub requested_agent_id: Option<AgentId>,
+    pub requested_project_id: Option<ProjectId>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ConversationBindingResolution {
+    pub tenant_id: TenantId,
+    pub actor: TurnActor,
+    pub turn_scope: TurnScope,
+    pub source_binding_ref: SourceBindingRef,
+    pub reply_target_binding_ref: ReplyTargetBindingRef,
+    pub access: ThreadAccessDecision,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct LinkConversationRequest {
+    pub tenant_id: TenantId,
+    pub adapter_kind: AdapterKind,
+    pub adapter_installation_id: AdapterInstallationId,
+    pub external_actor_ref: ExternalActorRef,
+    pub external_conversation_ref: ExternalConversationRef,
+    pub target_thread_id: ThreadId,
+    pub target_agent_id: Option<AgentId>,
+    pub target_project_id: Option<ProjectId>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct LinkedConversationBinding {
+    pub thread_id: ThreadId,
+    pub source_binding_ref: SourceBindingRef,
+    pub reply_target_binding_ref: ReplyTargetBindingRef,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ReplyTargetBinding {
+    pub tenant_id: TenantId,
+    pub actor_user_id: UserId,
+    pub thread_id: ThreadId,
+    pub external_conversation_ref: ExternalConversationRef,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ThreadAccessDecision {
+    Allowed,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum MessageIdempotencyStatus {
+    Inserted,
+    Duplicate,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct AcceptInboundMessageRequest {
+    pub tenant_id: TenantId,
+    pub thread_id: ThreadId,
+    pub actor: TurnActor,
+    pub source_binding_ref: SourceBindingRef,
+    pub reply_target_binding_ref: ReplyTargetBindingRef,
+    pub external_event_id: ExternalEventId,
+    pub content_ref: InboundMessageContentRef,
+    pub received_at: DateTime<Utc>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct AcceptedInboundMessage {
+    pub tenant_id: TenantId,
+    pub thread_id: ThreadId,
+    pub message_ref: AcceptedMessageRef,
+    pub source_binding_ref: SourceBindingRef,
+    pub reply_target_binding_ref: ReplyTargetBindingRef,
+    pub idempotency: MessageIdempotencyStatus,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ThreadMessageRecord {
+    pub accepted: AcceptedInboundMessage,
+    pub actor: TurnActor,
+    pub external_event_id: ExternalEventId,
+    pub content_ref: InboundMessageContentRef,
+    pub received_at: DateTime<Utc>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct InboundTurnRequest {
+    pub tenant_id: TenantId,
+    pub adapter_kind: AdapterKind,
+    pub adapter_installation_id: AdapterInstallationId,
+    pub external_actor_ref: ExternalActorRef,
+    pub external_conversation_ref: ExternalConversationRef,
+    pub external_event_id: ExternalEventId,
+    pub content_ref: InboundMessageContentRef,
+    pub requested_agent_id: Option<AgentId>,
+    pub requested_project_id: Option<ProjectId>,
+    pub received_at: DateTime<Utc>,
+    pub requested_run_profile: Option<RunProfileRequest>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct InboundTurnResponse {
+    pub resolution: ConversationBindingResolution,
+    pub accepted_message: AcceptedInboundMessage,
+    pub turn_submission: Option<SubmitTurnResponse>,
+}
